# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PlayZone Bar API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated

  # RESERVATIONS
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: upcoming
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # COUPONS
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List coupons
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create coupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/{id}:
    patch:
      operationId: updateCoupon
      tags: [coupons]
      summary: Update coupon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      summary: Delete coupon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponValidateInput"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidation"

  # AVAILABILITY PATTERNS
  /availability:
    get:
      operationId: listAvailabilityPatterns
      tags: [availability]
      summary: List availability patterns
      responses:
        "200":
          description: Patterns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailabilityPattern"
    post:
      operationId: createAvailabilityPattern
      tags: [availability]
      summary: Create availability pattern
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityPatternInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityPattern"
  /availability/{id}:
    patch:
      operationId: updateAvailabilityPattern
      tags: [availability]
      summary: Update availability pattern
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityPatternUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityPattern"
    delete:
      operationId: deleteAvailabilityPattern
      tags: [availability]
      summary: Delete availability pattern
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # STATS
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, username, name, role]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
...
    Event:
      type: object
      required: [id, title, date, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
        endDate:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        capacity:
          type: ["integer", "null"]
        location:
          type: ["string", "null"]
        featured:
          type: boolean
        createdAt:
          type: string
    EventInput:
      type: object
      required: [title, date]
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
        endDate:
          type: string
        imageUrl:
          type: string
        price:
          type: number
        capacity:
          type: integer
        location:
          type: string
        featured:
          type: boolean
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
        endDate:
          type: string
        imageUrl:
          type: string
        price:
          type: number
        capacity:
          type: integer
        location:
          type: string
        featured:
          type: boolean
    Coupon:
      type: object
      required: [id, code, discountType, discountValue, active, uses, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        discountType:
          type: string
          enum: [percent, fixed]
        discountValue:
          type: number
        validUntil:
          type: ["string", "null"]
        maxUses:
          type: ["integer", "null"]
        uses:
          type: integer
        active:
          type: boolean
        createdAt:
          type: string
    CouponInput:
      type: object
      required: [code, discountType, discountValue]
      properties:
        code:
          type: string
        discountType:
          type: string
          enum: [percent, fixed]
        discountValue:
          type: number
        validUntil:
          type: string
        maxUses:
          type: integer
        active:
          type: boolean
    CouponUpdate:
      type: object
      properties:
        code:
          type: string
        discountType:
          type: string
          enum: [percent, fixed]
        discountValue:
          type: number
        validUntil:
          type: string
        maxUses:
          type: integer
        active:
          type: boolean
    CouponValidateInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
    CouponValidation:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        coupon:
          $ref: "#/components/schemas/Coupon"
        message:
          type: ["string", "null"]
    AvailabilityPattern:
      type: object
      required: [id, dayOfWeek, type, slots]
      properties:
        id:
          type: integer
        dayOfWeek:
          type: integer
          minimum: 0
          maximum: 6
        type:
          type: string
          enum: [salao, gamer]
        slots:
          type: array
          items:
            type: string
        active:
          type: boolean
    AvailabilityPatternInput:
      type: object
      required: [dayOfWeek, type, slots]
      properties:
        dayOfWeek:
          type: integer
          minimum: 0
          maximum: 6
        type:
          type: string
          enum: [salao, gamer]
        slots:
          type: array
          items:
            type: string
        active:
          type: boolean
    AvailabilityPatternUpdate:
      type: object
      properties:
        dayOfWeek:
          type: integer
        type:
          type: string
          enum: [salao, gamer]
        slots:
          type: array
          items:
            type: string
        active:
          type: boolean
...