# lib/api-zod/src/generated/types/availabilityPatternUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PlayZone Bar API
 * OpenAPI spec version: 0.1.0
 */
import type { AvailabilityPatternUpdateType } from './availabilityPatternUpdateType';

export interface AvailabilityPatternUpdate {
  dayOfWeek?: number;
  type?: AvailabilityPatternUpdateType;
  slots?: string[];
  active?: boolean;
}